Added DynamicMessageQueue to support dynamic parameter (In Memory) to ignored passing and researilizing/serializing Request and Response message for Local Spider  to run on single Process.
 try with .Net 4.5.2
    Support Page Limit
    Fixed bug of MemoryUsage used Int datatype to be ulong (Crashed on machine 12GB RAM)

+++ src/DotnetSpider/Downloader/Entity/DownloaderAgent.cs
@@ -36,7 +36,7 @@ namespace DotnetSpider.Downloader.Entity
 		/// 总内存
 		/// </summary>
 		[Column("total_memory")]
-		public int TotalMemory { get; set; }
+		public ulong TotalMemory { get; set; }
 
 		/// <summary>
 		/// 上一次更新时间
